requirements.txt
# LLM & AI
langchain==0.1.0
langchain-community==0.0.13
ollama==0.1.6

# Vector Database
chromadb==0.4.22

# UI
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0

# Document Processing (เพิ่มตรงนี้!) 🔥
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.2